+++ src/main/java/com/tripj/domain/checklist/repository/CheckListRepositoryCustom.java
@@ -1,8 +1,7 @@
 package com.tripj.domain.checklist.repository;
 
-import com.tripj.domain.checklist.model.dto.GetCheckListResponse;
-import com.tripj.domain.checklist.model.dto.GetMyCheckListRequest;
-import com.tripj.domain.checklist.model.dto.GetMyCheckListResponse;
+import com.tripj.domain.checklist.model.dto.response.GetCheckListResponse;
+import com.tripj.domain.checklist.model.dto.response.GetMyCheckListResponse;
 
 import java.util.List;
