+++ src/main/java/com/tripj/domain/checklist/service/CheckListService.java
@@ -1,6 +1,8 @@
 package com.tripj.domain.checklist.service;
 
-import com.tripj.domain.checklist.model.dto.*;
+import com.tripj.domain.checklist.model.dto.request.CreateCheckListRequest;
+import com.tripj.domain.checklist.model.dto.request.PackCheckListRequest;
+import com.tripj.domain.checklist.model.dto.response.*;
 import com.tripj.domain.checklist.model.entity.CheckList;
 import com.tripj.domain.checklist.repository.CheckListRepository;
 import com.tripj.domain.item.model.entity.Item;
