+++ src/main/java/com/tripj/domain/item/service/ItemService.java
@@ -2,8 +2,8 @@ package com.tripj.domain.item.service;
 
 import com.tripj.domain.country.model.entity.Country;
 import com.tripj.domain.country.repository.CountryRepository;
-import com.tripj.domain.item.model.dto.CreateItemRequest;
-import com.tripj.domain.item.model.dto.CreateItemResponse;
+import com.tripj.domain.item.model.dto.request.CreateItemRequest;
+import com.tripj.domain.item.model.dto.response.CreateItemResponse;
 import com.tripj.domain.item.model.entity.Item;
 import com.tripj.domain.item.repository.ItemRepository;
 import com.tripj.domain.itemcate.model.entity.ItemCate;
