+++ src/main/java/com/tripj/domain/like/model/dto/response/CreateLikedBoardResponse.java
@@ -1,4 +1,4 @@
-package com.tripj.domain.like.model.dto;
+package com.tripj.domain.like.model.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
