+++ src/main/java/com/tripj/domain/trip/model/dto/response/CreateTripResponse.java
@@ -1,10 +1,9 @@
-package com.tripj.domain.trip.model.dto;
+package com.tripj.domain.trip.model.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
