+++ src/main/java/com/tripj/domain/trip/model/dto/response/GetTripResponse.java
@@ -1,4 +1,4 @@
-package com.tripj.domain.trip.model.dto;
+package com.tripj.domain.trip.model.dto.response;
 
 import com.querydsl.core.annotations.QueryProjection;
 import io.swagger.v3.oas.annotations.media.Schema;
