+++ src/main/java/com/tripj/domain/trip/model/dto/response/UpdateTripResponse.java
@@ -1,11 +1,9 @@
-package com.tripj.domain.trip.model.dto;
+package com.tripj.domain.trip.model.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-
-import java.time.LocalDate;
 
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
