+++ src/main/java/com/tripj/domain/trip/repository/TripRepositoryCustom.java
@@ -1,11 +1,9 @@
 package com.tripj.domain.trip.repository;
 
 
-import com.tripj.domain.trip.model.dto.GetTripResponse;
-import com.tripj.domain.trip.model.entity.Trip;
+import com.tripj.domain.trip.model.dto.response.GetTripResponse;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface TripRepositoryCustom {
